MTDSA-11940 New package structure
